Logging & Debugging implementations for curses UI.
* Logging in curses UI
* Curses UI can directly control solenoids
* Code cleanup & small fixes

Squashed commit of branch 'logging'.

Squashed commit of the following:

commit 5552ae0e7d4a564bff2df1e2854d3a2d8fd5ec7e
Author: Henning Hasemann <henning@hasemail.de>
Date:   Thu May 10 15:24:40 2018 +0200

    missing include for ringbuffer

commit b200fc3a2272f0d42efff5e4a4b27f3ea5c94a68
Author: Henning Hasemann <henning@hasemail.de>
Date:   Thu Mar 29 23:18:01 2018 +0200

    -Wall and some resulting fixes

commit ee49cc2252a3dd7fc760922eeaa7c95968140071
Author: Henning Hasemann <henning@hasemail.de>
Date:   Thu Mar 29 23:17:41 2018 +0200

    BufferLogger "scroll" FPS messages for curses interface

commit 2a4f0ad0d4f5c2ed6955b75172f2d028be4fb3cf
Author: Henning Hasemann <henning@hasemail.de>
Date:   Wed Mar 28 16:09:00 2018 +0200

    WIP

commit 9f0ff1cec7c9bfc1923836f17e77bfdcc9e09a92
Author: Henning Hasemann <henning@hasemail.de>
Date:   Wed Mar 28 13:25:53 2018 +0200

    Ball Return uses now AUX0, code cleanup

commit 76797cf2270ec3cd39e1ba0967f2f477b7b13913
Author: Henning Hasemann <henning@hasemail.de>
Date:   Wed Mar 28 11:42:40 2018 +0200

    Curses UI allows direct control of solenoids

+++ boards/control_module/software/src/pcf_font.cc
@@ -26,7 +26,7 @@ namespace {
 		is.read(reinterpret_cast<char*>(buffer), sizeof(buffer));
 
 		R r = 0;
-		for(int i = 0; i < sizeof(buffer); i++) {
+		for(size_t i = 0; i < sizeof(buffer); i++) {
 			r |= buffer[i] << OrderTrait::shift(i, sizeof(buffer));
 		}
 		return r;
@@ -108,7 +108,7 @@ void PcfFont::load_font(const char *filename) {
 	const int bytes_per_row = 1 << (bitmaps.format & PCF_GLYPH_PAD_MASK);
 	const float size_factor = 2.0 / bytes_per_row;
 
-	for(int glyph = 0; glyph < bitmaps.glyph_count - 1; glyph++) {
+	for(size_t glyph = 0; glyph < bitmaps.glyph_count - 1; glyph++) {
 		int offset_start = bitmaps.offsets[glyph];
 		int offset_end = bitmaps.offsets[glyph + 1];
 
@@ -222,10 +222,10 @@ PcfFont::Bitmaps::Bitmaps(std::ifstream& pcf)
 	format = read_lsb_u32(pcf);
 	glyph_count = read_u32(pcf, format);
 	offsets = new uint32_t[glyph_count];
-	for(int i = 0; i < glyph_count; i++) {
+	for(size_t i = 0; i < glyph_count; i++) {
 		offsets[i] = read_u32(pcf, format);
 	}
-	for(int i = 0; i < 4; i++) {
+	for(size_t i = 0; i < 4; i++) {
 		bitmap_sizes[i] = read_u32(pcf, format);
 	}
 	uint32_t sz = bitmap_sizes[format & PCF_GLYPH_PAD_MASK];
